Add output of frame delta to JSON.
This will allow programs that are designed to sync to 
the output of recordings. 

Syncing to recorded live gameplay sessions where 
the replay of a goal is still present will not want to 
use the delta field, while recordings of the replay 
afterwards will need to use the frame delta to 
approriatly react to the missing goal replay.